refactor(dashboard): extract websocket meeting chat to chat.rs (#1266)
routes.rs: 3219 -> 2881
chat.rs: NEW 346

Moves the WebSocket meeting facilitator chat module:
- load_dashboard_meeting_prompt
- open_dashboard_agent_session
- ws_chat_handler (now pub(crate))
- handle_ws_chat (now pub(crate))

Re-imports ws_chat_handler back into routes.rs for the router.
